requirements.txt
dependency_injector==4.41.0
Django==5.0.3
djangorestframework==3.15.1
drf_yasg==1.21.7
petl==1.7.15
pydantic==2.7.0
pytest==8.1.1
Requests==2.31.0
